Decouple `LoginScreen` from `RootScreen`
